Created validation for server settings schema
